notebooks/nws_update_gcf/requirements.txt
datetime
pandas
numpy
requests
beautifulsoup4
google-cloud-bigquery
google-cloud-storage
google-cloud-logging
# google-auth -- Unneeded for GCF
# google-auth-oauthlib
# google-auth-httplib2
